# my-SMS-app/apps/edge/api/internal/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Edge API (SMS Dev)
  version: 0.1.0
servers:
  - url: http://localhost:8081
paths:
...
  /logbooks/{type}:
    get:
      summary: List logbook entries by type + tenant + vessel
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string, enum: [bridge, engine] }
        - in: query
          name: tenantId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: vesselId
          required: true
          schema: { type: string, format: uuid }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LogbookEntry' }
    post:
      summary: Create a logbook entry
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string, enum: [bridge, engine] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tenantId: { type: string, format: uuid }
                vesselId: { type: string, format: uuid }
                data: { type: object }
              required: [tenantId, vesselId, data]
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LogbookEntry' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LogbookEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        vesselId: { type: string, format: uuid }
        type: { type: string, enum: [bridge, engine] }
        ts: { type: string, format: date-time }
        data: { type: object }
        correctionRequested: { type: boolean }
        correctionReason: { type: string, nullable: true }
        correctionBy: { type: string, nullable: true }
        correctionAt: { type: string, format: date-time, nullable: true }
        countersignedBy: { type: string, nullable: true }
        countersignedAt: { type: string, format: date-time, nullable: true }